requirements.txt
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-ollama==0.0.1
